Create RenameSDX.c
Written to speed up testing.  Removes part of a file name so that the file can be taken back in to the test system.
Currently will remove the last seven characters from the file name.
This is used to remove the time stamp from the end of the file name on files.